src-cuda/requirements.txt
huggingface
transformers==4.45.1
Pillow
numpy
einops
tqdm
dataclasses
timm
torchvision